feat: unique duplicate logic
